rename: header-buttons -> header-actions
